requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
requests==2.31.0
certifi==2023.11.17
urllib3==2.1.0
sqlite3-binary